Moved model to its own directory
