+++ src/model/loader/fbx/data/texture.rs
