+++ src/model/loader/obj.rs
